backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - PostgreSQL & MySQL Support
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.9
cryptography==41.0.7

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# Data validation
pydantic[email]==2.5.0
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0

# NEW: Cloud Storage (AWS S3, Supabase)
boto3==1.34.0
botocore==1.34.0
supabase==2.3.0

# NEW: OAuth Authentication (Google, GitHub)
authlib==1.3.0
itsdangerous==2.1.2
requests==2.31.0

# NEW: OCR (Optical Character Recognition)
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78

# AI/ML (Optional - for AI features)
# transformers==4.35.2
# torch==2.1.1
# scikit-learn==1.3.2

